refactor: mutation 함수 분리
